+++ app/join/step5/page.tsx
@@ -1,22 +1,9 @@
-'use client';
-import { useState } from 'react';
-
-import Button from '@/components/common/Button';
 import TopNavigationBar from '@/components/common/NavigationBar/TopNavigationBar';
-import PersonalityItem from '@/components/join/PersonalityItem';
 import { AuthTitleTextMessage } from '@/components/TextMessage/AuthTextMessage';
-import { personalityList } from '@/constants/personalityList';
+
+import InputForm from './InputForm';
 
 export default function Step5Page() {
-  const [selectedPersonalityList, setSelectedPersonalityList] = useState<number[]>([]);
-  const handlePersonalityClick = (index: number) => {
-    if (selectedPersonalityList.includes(index)) {
-      setSelectedPersonalityList(selectedPersonalityList.filter((item) => item !== index));
-    } else {
-      setSelectedPersonalityList([...selectedPersonalityList, index]);
-    }
-  };
-
   return (
     <div className="relative h-full ">
       <TopNavigationBar text="회원가입" />
@@ -27,24 +14,9 @@ export default function Step5Page() {
 
       <p className="text-14 text-gray2">3개를 선택해주세요.</p>
 
+      <InputForm />
+
       <div className="h-30" />
-
-      <section className="flex flex-wrap gap-12">
-        {personalityList.map((personality, index: number) => (
-          <PersonalityItem
-            key={personality.id}
-            personality={personality}
-            isSelected={selectedPersonalityList.includes(index)}
-            onClick={() => {
-              handlePersonalityClick(index);
-            }}
-          />
-        ))}
-      </section>
-
-      <section className="absolute bottom-0 w-full">
-        <Button text="완료" disabled={selectedPersonalityList.length === 0} href="/login" />
-      </section>
     </div>
   );
 }
